feat: TTShaderTextureUsageInformationRegistry
