+++ Runtime/TextureAtlas/AAOCode/ShaderInformation.Liltoon.cs
@@ -6,36 +6,34 @@ namespace net.rs64.TexTransTool.TextureAtlas.AAOCode
 {
     internal class LiltoonShaderInformationWithOutline : ITTShaderTextureUsageInformation
     {
-
-        public int Priority => 0;
-
         public void GetMaterialTextureUVUsage(ITTTextureUVUsageWriter writer)
         {
             LiltoonShaderInformation.GetMaterialInformation(new MaterialInformationTranslator(writer), true);
         }
-
-        public bool IsSupportShader(Shader shader)
-        {
-            return shader.name.Contains("lilToon") && shader.name.Contains("Outline");
-        }
     }
     internal class LiltoonShaderInformationWithOutOutline : ITTShaderTextureUsageInformation
     {
-
-        public int Priority => 0;
-
         public void GetMaterialTextureUVUsage(ITTTextureUVUsageWriter writer)
         {
             LiltoonShaderInformation.GetMaterialInformation(new MaterialInformationTranslator(writer), false);
         }
-
-        public bool IsSupportShader(Shader shader)
-        {
-            return shader.name.Contains("lilToon") && shader.name.Contains("Outline") is false;
-        }
     }
     internal class LiltoonShaderInformation
     {
+        [TexTransCoreEngineForUnity.TexTransInitialize]
+        internal static void liltoonShaderInformationRegistering()
+        {
+            var lilOutLine = new LiltoonShaderInformationWithOutline();
+            var lil = new LiltoonShaderInformationWithOutOutline();
+            foreach (var shader in TexTransCoreEngineForUnity.TexTransCoreRuntime.LoadAssetsAtType(typeof(Shader)).Cast<Shader>())
+            {
+                if (shader.name.Contains("lilToon"))
+                {
+                    if (shader.name.Contains("Outline")) { TTShaderTextureUsageInformationRegistry.RegisterTTShaderTextureUsageInformation(shader, lilOutLine); }
+                    else { TTShaderTextureUsageInformationRegistry.RegisterTTShaderTextureUsageInformation(shader, lil); }
+                }
+            }
+        }
 
         public static void GetMaterialInformation(IMaterialInformationCallbackAbstractionInterface matInfo, bool isOutline)
         {
