+++ Runtime/TextureAtlas/ShaderTextureUsageInformation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ec44c47edb46e753891916acdec36bcf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
