+++ Runtime/TextureAtlas/ShaderTextureUsageInformation/ITTShaderTextureUsageInformation.cs.meta
