+++ Runtime/TextureAtlas/ShaderTextureUsageInformation/MaterialInformationTranslator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d74e4f943ee3a8c78c8cd94092c6e17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
