+++ Runtime/TextureAtlas/ShaderTextureUsageInformation/TTShaderTextureUsageInformationRegistry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9478db9f47a07d2cebea8d7025bc8d59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
